Rename tableview cells files, remove main model and defaul-image files
